refactor(close): refactor close into an Operation
