+++ lib/mongo_client.js
@@ -4,7 +4,6 @@ const ChangeStream = require('./change_stream');
 const Db = require('./db');
 const EventEmitter = require('events').EventEmitter;
 const executeOperation = require('./operations/execute_operation');
-const executeLegacyOperation = require('./utils').executeLegacyOperation;
 const inherits = require('util').inherits;
 const MongoError = require('./core').MongoError;
 const deprecate = require('util').deprecate;
@@ -12,7 +11,7 @@ const WriteConcern = require('./write_concern');
 
 // Operations
 const ConnectOperation = require('./operations/connect');
-const closeOperation = require('./operations/mongo_client_ops').closeOperation;
+const CloseOperation = require('./operations/close');
 
 /**
  * @fileOverview The **MongoClient** class is a class that allows for making Connections to MongoDB.
@@ -215,9 +214,8 @@ MongoClient.prototype.logout = deprecate(function(options, callback) {
  */
 MongoClient.prototype.close = function(force, callback) {
   if (typeof force === 'function') (callback = force), (force = false);
-  return executeLegacyOperation(this, closeOperation, [this, force, callback], {
-    skipSessions: true
-  });
+  const operation = new CloseOperation(this, force);
+  return executeOperation(this, operation, callback);
 };
 
 /**
